# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/landing-mode": {
      "post": {
        "operationId": "LandingModeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLandingModeRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLandingModeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "LandingMode"
        ]
      },
      "get": {
        "operationId": "LandingModeController_readAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadLandingModeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "LandingMode"
        ]
      }
    },
    "/landing-mode/readById/{id}": {
      "get": {
        "operationId": "LandingModeController_readById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "LandingMode"
        ]
      }
    },
    "/landing-mode/{id}": {
      "patch": {
        "operationId": "LandingModeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLandingModeRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLandingModeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "LandingMode"
        ]
      },
      "delete": {
        "operationId": "LandingModeController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLandingModeResponseDto"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "LandingMode"
        ]
      }
    },
...
    "/templates/landing-mode/{landingModeId}": {
      "get": {
        "operationId": "TemplateController_getTemplatesByLandingModeId",
        "parameters": [
          {
            "name": "landingModeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadTemplateResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "Template"
        ]
      }
    },
...
    "/contact-us/create": {
      "post": {
        "operationId": "ContactUsController_createContactUs",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactUsRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateContactUsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "ContactUs"
        ]
      }
    },
    "/contact-us/get": {
      "get": {
        "operationId": "ContactUsController_getContactUs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadContactUsResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "ContactUs"
        ]
      }
    },
    "/contact-us/update": {
      "patch": {
        "operationId": "ContactUsController_updateContactUs",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContactUsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateContactUsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "ContactUs"
        ]
      }
    },
    "/contact-us/upsert": {
      "post": {
        "operationId": "ContactUsController_upsertContactUs",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactUsRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateContactUsResponseDto"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "ContactUs"
        ]
      }
    },
    "/complaints": {
      "post": {
        "operationId": "ComplaintController_createComplaint",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComplaintRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateComplaintResponseDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "summary": "Submit a complaint, suggestion, or message",
        "tags": [
          "Complaint"
        ]
      },
      "get": {
        "operationId": "ComplaintController_getAllComplaints",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadComplaintResponseDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        },
        "summary": "Get all complaints (Admin)",
        "tags": [
          "Complaint"
        ]
      }
    },
    "/complaints/readById/{id}": {
      "get": {
        "operationId": "ComplaintController_getComplaintById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadComplaintResponseDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "summary": "Get a complaint by ID",
        "tags": [
          "Complaint"
        ]
      }
    },
    "/complaints/my-complaints": {
      "get": {
        "operationId": "ComplaintController_getMyComplaints",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadComplaintResponseDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        },
        "summary": "Get all complaints submitted by the current authenticated user",
        "tags": [
          "Complaint"
        ]
      }
    },
...
  },
  "info": {
    "title": "rememeber-me",
    "description": "API documentation for rememeber-me.env",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateLandingModeRequestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the landing mode"
          },
          "description": {
            "type": "string",
            "description": "Description of the landing mode"
          },
          "iconId": {
            "type": "string",
            "description": "Icon file ID reference"
          },
          "landingModeType": {
            "type": "string",
            "description": "Type of the landing mode",
            "enum": [
              "full-mode",
              "event-mode",
              "video-only-mode"
            ],
            "example": "full-mode"
          },
          "iconURL": {
            "type": "string",
            "description": "Icon URL for display"
          },
          "isActive": {
            "type": "boolean",
            "description": "Status of landing mode (active/inactive)"
          }
        }
      },
      "CreateLandingModeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "ReadLandingModeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-landing-mode"
          },
          "title": {
            "type": "string",
            "example": "Dark Mode"
          },
          "description": {
            "type": "string",
            "example": "Landing mode for dark theme"
          },
          "landingModeType": {
            "type": "string",
            "example": "full-mode",
            "enum": [
              "full-mode",
              "event-mode",
              "video-only-mode"
            ]
          },
          "iconId": {
            "type": "string",
            "example": "icon-file-id"
          },
          "iconURL": {
            "type": "string",
            "example": "https://cdn.example.com/icons/darkmode.png"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          }
        },
        "required": [
          "id",
          "title",
          "landingModeType",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateLandingModeRequestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the landing mode"
          },
          "description": {
            "type": "string",
            "description": "Description of the landing mode"
          },
          "iconId": {
            "type": "string",
            "description": "Icon file ID reference"
          },
          "iconURL": {
            "type": "string",
            "description": "Icon URL for display"
          },
          "isActive": {
            "type": "boolean",
            "description": "Status of landing mode (active/inactive)"
          }
        }
      },
...
      "ReadTemplateResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-template"
          },
          "name": {
            "type": "string",
            "example": "Memorial Template A"
          },
          "description": {
            "type": "string",
            "example": "This template is used for modern memorials"
          },
          "landingModeId": {
            "type": "string",
            "example": "uuid-of-landing-mode"
          },
          "landingMode": {
            "$ref": "#/components/schemas/ReadLandingModeResponseDto"
          },
          "thumbnailURL": {
            "type": "string",
            "example": "https://cdn.example.com/templates/template-thumbnail.png"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateContactUsRequestDto": {
        "type": "object",
        "properties": {
          "addressLine1": {
            "type": "string",
            "example": "2118 Thornridge Cir."
          },
          "addressLine2": {
            "type": "string",
            "example": "Suite 100"
          },
          "city": {
            "type": "string",
            "example": "Syracuse"
          },
          "state": {
            "type": "string",
            "example": "Connecticut"
          },
          "zipCode": {
            "type": "string",
            "example": "35624"
          },
          "country": {
            "type": "string",
            "example": "United States"
          },
          "phoneNumber": {
            "type": "string",
            "example": "(171) 555-2111"
          },
          "emailAddress": {
            "type": "string",
            "example": "rememberme@demo.com"
          },
          "facebookUrl": {
            "type": "string",
            "example": "https://www.facebook.com/rememberme"
          },
          "instagramUrl": {
            "type": "string",
            "example": "https://www.instagram.com/rememberme"
          },
          "twitterUrl": {
            "type": "string",
            "example": "https://www.twitter.com/rememberme"
          },
          "linkedinUrl": {
            "type": "string",
            "example": "https://www.linkedin.com/company/rememberme"
          }
        }
      },
      "CreateContactUsResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-contact-us"
          }
        },
        "required": [
          "id"
        ]
      },
      "ReadContactUsResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-contact-us"
          },
          "addressLine1": {
            "type": "string",
            "example": "2118 Thornridge Cir."
          },
          "addressLine2": {
            "type": "string",
            "example": "Suite 100"
          },
          "city": {
            "type": "string",
            "example": "Syracuse"
          },
          "state": {
            "type": "string",
            "example": "Connecticut"
          },
          "zipCode": {
            "type": "string",
            "example": "35624"
          },
          "country": {
            "type": "string",
            "example": "United States"
          },
          "phoneNumber": {
            "type": "string",
            "example": "(171) 555-2111"
          },
          "emailAddress": {
            "type": "string",
            "example": "rememberme@demo.com"
          },
          "facebookUrl": {
            "type": "string",
            "example": "https://www.facebook.com/rememberme"
          },
          "instagramUrl": {
            "type": "string",
            "example": "https://www.instagram.com/rememberme"
          },
          "twitterUrl": {
            "type": "string",
            "example": "https://www.twitter.com/rememberme"
          },
          "linkedinUrl": {
            "type": "string",
            "example": "https://www.linkedin.com/company/rememberme"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-08-28T12:34:56Z"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateContactUsRequestDto": {
        "type": "object",
        "properties": {
          "addressLine1": {
            "type": "string",
            "example": "2118 Thornridge Cir."
          },
          "addressLine2": {
            "type": "string",
            "example": "Suite 100"
          },
          "city": {
            "type": "string",
            "example": "Syracuse"
          },
          "state": {
            "type": "string",
            "example": "Connecticut"
          },
          "zipCode": {
            "type": "string",
            "example": "35624"
          },
          "country": {
            "type": "string",
            "example": "United States"
          },
          "phoneNumber": {
            "type": "string",
            "example": "(171) 555-2111"
          },
          "emailAddress": {
            "type": "string",
            "example": "rememberme@demo.com"
          },
          "facebookUrl": {
            "type": "string",
            "example": "https://www.facebook.com/rememberme"
          },
          "instagramUrl": {
            "type": "string",
            "example": "https://www.instagram.com/rememberme"
          },
          "twitterUrl": {
            "type": "string",
            "example": "https://www.twitter.com/rememberme"
          },
          "linkedinUrl": {
            "type": "string",
            "example": "https://www.linkedin.com/company/rememberme"
          }
        }
      },
      "UpdateContactUsResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-contact-us"
          }
        },
        "required": [
          "id"
        ]
      },
      "CreateComplaintRequestDto": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "example": "Perry Wilson"
          },
          "userEmail": {
            "type": "string",
            "example": "perry.wilson@example.com"
          },
          "userPhoneNumber": {
            "type": "string",
            "example": "(219) 555-0114"
          },
          "messageType": {
            "type": "string",
            "example": "message",
            "enum": [
              "complaint",
              "suggestion",
              "message"
            ],
            "description": "Type of message: complaint, suggestion, or message",
            "default": "message"
          },
          "subject": {
            "type": "string",
            "example": "Question about memorial services"
          },
          "messageContent": {
            "type": "string",
            "example": "I would like to know more about your memorial services..."
          }
        },
        "required": [
          "userName",
          "userEmail",
          "messageContent"
        ]
      },
      "CreateComplaintResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-complaint"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}